+++ src/test/java/com/moabam/api/domain/room/CertificationTest.java
@@ -1,14 +1,9 @@
-package com.moabam.api.domain.entity;
+package com.moabam.api.domain.room;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import com.moabam.api.domain.room.Certification;
-import com.moabam.api.domain.room.Room;
-import com.moabam.api.domain.room.RoomType;
-import com.moabam.api.domain.room.Routine;
 
 class CertificationTest {
